refactor: extract createFieldsStore
