+++ src/createForm.js
@@ -3,20 +3,20 @@ import createBaseForm from './createBaseForm';
 export const mixin = {
   getForm() {
     return {
-      getFieldsValue: this.getFieldsValue,
-      getFieldValue: this.getFieldValue,
+      getFieldsValue: this.fieldsStore.getFieldsValue,
+      getFieldValue: this.fieldsStore.getFieldValue,
       getFieldInstance: this.getFieldInstance,
       setFieldsValue: this.setFieldsValue,
       setFields: this.setFields,
-      setFieldsInitialValue: this.setFieldsInitialValue,
+      setFieldsInitialValue: this.fieldsStore.setFieldsInitialValue,
       getFieldDecorator: this.getFieldDecorator,
       getFieldProps: this.getFieldProps,
-      getFieldsError: this.getFieldsError,
-      getFieldError: this.getFieldError,
-      isFieldValidating: this.isFieldValidating,
-      isFieldsValidating: this.isFieldsValidating,
-      isFieldsTouched: this.isFieldsTouched,
-      isFieldTouched: this.isFieldTouched,
+      getFieldsError: this.fieldsStore.getFieldsError,
+      getFieldError: this.fieldsStore.getFieldError,
+      isFieldValidating: this.fieldsStore.isFieldValidating,
+      isFieldsValidating: this.fieldsStore.isFieldsValidating,
+      isFieldsTouched: this.fieldsStore.isFieldsTouched,
+      isFieldTouched: this.fieldsStore.isFieldTouched,
       isSubmitting: this.isSubmitting,
       submit: this.submit,
       validateFields: this.validateFields,
